refactor: migrer imports fra localServiceClient til direkteklienter (#164)
Migrerer 18 av 35 filer fra barrel-barrel til direkte domeneklient-importer.
Markerer localServiceClient.ts som deprecated med JSDoc-kommentar.
Oppdaterer tilhørende testfiler med korrekte vi.mock-målpunkter.
